MS: Add CRTFragment.
This is a functional overlay class for the Cosmic Ray
Tagger/Tracker.  It provides functionality for decoding
raw CRT data and checking its integrity.

Future commits will update it to the newer data scheme
that hides the fact that the CRT internally only has
a 32-bit clock counter.  References to the Unix time
in this version can be disregarded.